+++ src/components/core/Footer.vue
@@ -1,0 +1,11 @@
+<template>
+  <v-footer app>
+    <span>&copy; Esteban y Miguel Castro 2019</span>
+  </v-footer>
+</template>
+
+<script>
+export default {
+  name: 'Footer',
+};
+</script>