+++ src/plugins/common.js
@@ -1,1 +1,4 @@
 import Vue from 'vue';
+import ErrorMessage from '@/components/common/ErrorMessage.vue';
+
+Vue.component('ErrorMessage', ErrorMessage);