+++ CHANGELOG.md
@@ -1,8 +1,57 @@
+## v0.0.14 (November 07, 2019)
+
+- Add refreshToken action in auth.js
+- Add register api / module / component
+- Add logout button in toolbar
+- Add refreshToken user function
+
+## v0.0.13 (November 03, 2019)
+
+- Add date-fns package
+- Add Auth store
+- Add Auth api module
+- Add env.development variables
+- Add utils and updates files
+- Finish Login api call
+- Logout action and mutation
+- Login action and mutation
+- Save token action and mutation
+- Loading component and store module
+- Error component and store module
+
+## v0.0.12 (November 02, 2019)
+
+- Update Login View/component
+- Add form login validations
+- Create function form login submit
+- Change default lang in veevalidate storage
+
+## v0.0.11 (November 01, 2019)
+
+- Add Login Component/view and link in Login Button
+- Refactor menuItem computed in Navigation Drawer
+
+## v0.0.10 (September 30, 2019)
+
+- Update button Toolbar
+- Clean function button collapse Navigation Drawer
+- Add login button to Toolbar
+- Add search bar to Toolbar
+- Add Home Button to Navigation drawer
+
+## v0.0.9 (September 22, 2019)
+
+- Add Toolbar Component
+- Add Navigation Drawer Component
+- Add Footer Component
+- Add Module navigationDreawer in Storage
+
 ## v0.0.8 (September 21, 2019)
 
 - Add axios
 - Add vue-cli-axios
 - Add common plugin
+- Add Poc layout
 
 ## v0.0.7 (September 10, 2019)